Replace uses of deprecated std::random_shuffle
It appears `std::random_shuffle` was deprecated in C++14 and removed in
C++17. It was replaced with `std::shuffle`, which is very similar but
takes a (required) explicit RNG parameter.

I've replaced these calls in the most trivial possible way: initializing
a Mersenne twister RNG at the top of the `sgd` function (in both cases)
and using it for the `shuffle` call. There may be performance reasons to
initialize the RNG elsewhere, but this at least seems to "work," so
there's that.